+++ client/index.ts
@@ -29,7 +29,7 @@ import { generateHS256JWS, generateRS256JWS } from "./src/utils/generateJWS";
 import { rsaPublicKeyToPEM } from "./src/utils/rsaPublicKeyToPEM";
 
 class OAuth2Client {
-  readonly openIDConfigurationURL: string;
+  readonly providerURL: string;
   readonly clientID: string;
   readonly clientSecret: string;
   readonly redirectURI: string;
@@ -39,7 +39,7 @@ class OAuth2Client {
   openIDConfiguration?: OpenIDConfiguration;
 
   constructor({
-    openIDConfigurationURL,
+    providerURL,
     clientID,
     clientSecret,
     redirectURI,
@@ -47,7 +47,7 @@ class OAuth2Client {
     scopes,
     fetch,
   }: OAuth2ClientConstructor) {
-    this.openIDConfigurationURL = openIDConfigurationURL;
+    this.providerURL = providerURL;
     this.clientID = clientID;
     this.clientSecret = clientSecret;
     this.redirectURI = redirectURI;
@@ -60,7 +60,13 @@ class OAuth2Client {
     if (this.openIDConfiguration) {
       return Promise.resolve();
     } else {
-      await this.fetch(this.openIDConfigurationURL)
+      const route = ".well-known/openid-configuration";
+      const openIDConfigurationURL = new URL(
+        route,
+        this.providerURL,
+      ).toString();
+
+      await this.fetch(openIDConfigurationURL)
         .then((response) => response.json())
         .then((openIDConfiguration) => {
           this.openIDConfiguration = openIDConfiguration;
@@ -248,10 +254,7 @@ class OAuth2Client {
     }
   }
 
-  async refreshTokens(
-    refresh_token: string,
-    providerUrl: string,
-  ): Promise<RefreshTokenResponse> {
+  async refreshTokens(refresh_token: string): Promise<RefreshTokenResponse> {
     await this.setOpenIDConfiguration();
 
     const payload = {
@@ -263,7 +266,7 @@ class OAuth2Client {
     };
 
     const route = "oauth/token";
-    const absoluteURL = new URL(route, providerUrl).toString();
+    const absoluteURL = new URL(route, this.providerURL).toString();
 
     return this.fetch(absoluteURL, {
       method: "POST",